requirements.txt
Flask==3.0.3
jinja2==3.1.3
gunicorn==21.2.0
python-dotenv==1.0.1
google-generativeai==0.4.1
openai==0.28